Code for Finite Differences for Polynonmials
This is a set of Subs/Macros that I created to help a student with finite differences in order to show how to populate the table as well as to determine the leading coefficient and degree from the table. I should be noted that if it is going to be used, it requires an ActiveX button called 'StartTestButton' or the calling of the sub 'Move_Button' should be commented out/removed when running the sub 'Run_Tester'.